Renew pure prompt setup
The old setup including it as a "theme" does no longer fit the current
code of the project.

This removes the zsh/themes directory (and pure as a submodule in there)
and instead adds pure as a submodule to the zsh/plugins directory.

The prompt then requires to functions to be symlinked / available in
$fpath. I therefore added a separate zsh/functions directory that I then
manually add to the $fpath in the zshrc.

The functions directory contains two symlinks for the functions to set
up the pure promt. The promt itself is then initialized in the zshrc
itself.